New API functions: `.isConnected()` `.isStreaming()`
All checks for connection state are normalized to use it.

The name `.connected()` may have been more consistent with the rest of
the library, but it would incorrectly evaluate to true in code that
used the old, undocumented `.connected` field, generating hard-to-find
bugs.  So `isConnected()` was chosen to avoid that situation.